Move logic into respective vault types
