+++ radix-engine/src/blueprints/resource/non_fungible/non_fungible_vault.rs
@@ -5,7 +5,7 @@ use crate::kernel::heap::{DroppedBucket, DroppedBucketResource};
 use crate::kernel::kernel_api::{KernelNodeApi, KernelSubstateApi};
 use crate::types::*;
 use native_sdk::runtime::Runtime;
-use radix_engine_interface::api::ClientApi;
+use radix_engine_interface::api::{ClientApi, ClientSubstateApi, LockFlags};
 use radix_engine_interface::api::types::*;
 use radix_engine_interface::blueprints::resource::*;
 
@@ -331,5 +331,297 @@ impl NonFungibleVaultBlueprint {
     {
         let info = VaultInfoSubstate::of(receiver, api)?;
         Ok(info.resource_address)
+    }
+}
+
+pub struct NonFungibleVault;
+
+impl NonFungibleVault {
+    pub fn liquid_amount<Y>(receiver: &RENodeId, api: &mut Y) -> Result<Decimal, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientSubstateApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LiquidNonFungible),
+            LockFlags::read_only(),
+        )?;
+        let substate_ref: &LiquidNonFungibleResource = api.kernel_get_substate_ref(handle)?;
+        let amount = substate_ref.amount();
+        api.sys_drop_lock(handle)?;
+        Ok(amount)
+    }
+
+    pub fn locked_amount<Y>(receiver: &RENodeId, api: &mut Y) -> Result<Decimal, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientSubstateApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LockedNonFungible),
+            LockFlags::read_only(),
+        )?;
+        let substate_ref: &LockedNonFungibleResource = api.kernel_get_substate_ref(handle)?;
+        let amount = substate_ref.amount();
+        api.sys_drop_lock(handle)?;
+        Ok(amount)
+    }
+
+    pub fn is_locked<Y>(receiver: &RENodeId, api: &mut Y) -> Result<bool, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientSubstateApi<RuntimeError>,
+    {
+        Ok(!Self::locked_amount(receiver, api)?.is_zero())
+    }
+
+    pub fn liquid_non_fungible_local_ids<Y>(
+        receiver: &RENodeId,
+        api: &mut Y,
+    ) -> Result<BTreeSet<NonFungibleLocalId>, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientSubstateApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LiquidNonFungible),
+            LockFlags::read_only(),
+        )?;
+        let substate_ref: &LiquidNonFungibleResource = api.kernel_get_substate_ref(handle)?;
+        let ids = substate_ref.ids().clone();
+        api.sys_drop_lock(handle)?;
+        Ok(ids)
+    }
+
+    pub fn locked_non_fungible_local_ids<Y>(
+        receiver: &RENodeId,
+        api: &mut Y,
+    ) -> Result<BTreeSet<NonFungibleLocalId>, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientSubstateApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LockedNonFungible),
+            LockFlags::read_only(),
+        )?;
+        let substate_ref: &LockedNonFungibleResource = api.kernel_get_substate_ref(handle)?;
+        let ids = substate_ref.ids();
+        api.sys_drop_lock(handle)?;
+        Ok(ids)
+    }
+
+    pub fn take<Y>(
+        receiver: &RENodeId,
+        amount: Decimal,
+        api: &mut Y,
+    ) -> Result<LiquidNonFungibleResource, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LiquidNonFungible),
+            LockFlags::MUTABLE,
+        )?;
+        let substate_ref: &mut LiquidNonFungibleResource =
+            api.kernel_get_substate_ref_mut(handle)?;
+        let taken = substate_ref.take_by_amount(amount).map_err(|e| {
+            RuntimeError::ApplicationError(ApplicationError::VaultError(VaultError::ResourceError(
+                e,
+            )))
+        })?;
+        api.sys_drop_lock(handle)?;
+
+        Runtime::emit_event(api, WithdrawResourceEvent::Amount(amount))?;
+
+        Ok(taken)
+    }
+
+    pub fn take_non_fungibles<Y>(
+        receiver: &RENodeId,
+        ids: &BTreeSet<NonFungibleLocalId>,
+        api: &mut Y,
+    ) -> Result<LiquidNonFungibleResource, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LiquidNonFungible),
+            LockFlags::MUTABLE,
+        )?;
+        let substate_ref: &mut LiquidNonFungibleResource =
+            api.kernel_get_substate_ref_mut(handle)?;
+        let taken = substate_ref
+            .take_by_ids(ids)
+            .map_err(VaultError::ResourceError)
+            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::VaultError(e)))?;
+        api.sys_drop_lock(handle)?;
+
+        Runtime::emit_event(api, WithdrawResourceEvent::Ids(ids.clone()))?;
+
+        Ok(taken)
+    }
+
+    pub fn put<Y>(
+        receiver: &RENodeId,
+        resource: LiquidNonFungibleResource,
+        api: &mut Y,
+    ) -> Result<(), RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
+    {
+        if resource.is_empty() {
+            return Ok(());
+        }
+
+        let event = DepositResourceEvent::Ids(resource.ids().clone());
+
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LiquidNonFungible),
+            LockFlags::MUTABLE,
+        )?;
+        let substate_ref: &mut LiquidNonFungibleResource =
+            api.kernel_get_substate_ref_mut(handle)?;
+        substate_ref.put(resource).map_err(|e| {
+            RuntimeError::ApplicationError(ApplicationError::VaultError(VaultError::ResourceError(
+                e,
+            )))
+        })?;
+        api.sys_drop_lock(handle)?;
+
+        Runtime::emit_event(api, event)?;
+
+        Ok(())
+    }
+
+    // protected method
+    pub fn lock_amount<Y>(
+        receiver: &RENodeId,
+        amount: Decimal,
+        api: &mut Y,
+    ) -> Result<NonFungibleProof, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LockedNonFungible),
+            LockFlags::MUTABLE,
+        )?;
+        let mut locked: &mut LockedNonFungibleResource = api.kernel_get_substate_ref_mut(handle)?;
+        let max_locked: Decimal = locked.ids.len().into();
+
+        // Take from liquid if needed
+        if amount > max_locked {
+            let delta = amount - max_locked;
+            let resource = NonFungibleVault::take(receiver, delta, api)?;
+
+            locked = api.kernel_get_substate_ref_mut(handle)?; // grab ref again
+            for nf in resource.into_ids() {
+                locked.ids.insert(nf, 0);
+            }
+        }
+
+        // Increase lock count
+        let n: usize = amount
+            .to_string()
+            .parse()
+            .expect("Failed to convert amount to usize");
+        let ids_for_proof: BTreeSet<NonFungibleLocalId> =
+            locked.ids.keys().cloned().into_iter().take(n).collect();
+        for id in &ids_for_proof {
+            locked.ids.entry(id.clone()).or_default().add_assign(1);
+        }
+
+        // Issue proof
+        Ok(NonFungibleProof::new(
+            ids_for_proof.clone(),
+            btreemap!(
+                LocalRef::Vault(receiver.clone().into()) => ids_for_proof
+            ),
+        )
+            .map_err(|e| {
+                RuntimeError::ApplicationError(ApplicationError::VaultError(VaultError::ProofError(e)))
+            })?)
+    }
+
+    // protected method
+    pub fn lock_non_fungibles<Y>(
+        receiver: &RENodeId,
+        ids: BTreeSet<NonFungibleLocalId>,
+        api: &mut Y,
+    ) -> Result<NonFungibleProof, RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LockedNonFungible),
+            LockFlags::MUTABLE,
+        )?;
+        let mut locked: &mut LockedNonFungibleResource = api.kernel_get_substate_ref_mut(handle)?;
+
+        // Take from liquid if needed
+        let delta: BTreeSet<NonFungibleLocalId> = ids
+            .iter()
+            .cloned()
+            .filter(|id| !locked.ids.contains_key(id))
+            .collect();
+        NonFungibleVault::take_non_fungibles(receiver, &delta, api)?;
+
+        // Increase lock count
+        locked = api.kernel_get_substate_ref_mut(handle)?; // grab ref again
+        for id in &ids {
+            locked.ids.entry(id.clone()).or_default().add_assign(1);
+        }
+
+        // Issue proof
+        Ok(NonFungibleProof::new(
+            ids.clone(),
+            btreemap!(
+                LocalRef::Vault(receiver.clone().into()) => ids
+            ),
+        )
+            .map_err(|e| {
+                RuntimeError::ApplicationError(ApplicationError::VaultError(VaultError::ProofError(e)))
+            })?)
+    }
+
+    // protected method
+    pub fn unlock_non_fungibles<Y>(
+        receiver: &RENodeId,
+        ids: BTreeSet<NonFungibleLocalId>,
+        api: &mut Y,
+    ) -> Result<(), RuntimeError>
+        where
+            Y: KernelNodeApi + KernelSubstateApi + ClientApi<RuntimeError>,
+    {
+        let handle = api.sys_lock_substate(
+            receiver.clone(),
+            SubstateOffset::Vault(VaultOffset::LockedNonFungible),
+            LockFlags::MUTABLE,
+        )?;
+        let locked: &mut LockedNonFungibleResource = api.kernel_get_substate_ref_mut(handle)?;
+
+        let mut liquid_non_fungibles = BTreeSet::<NonFungibleLocalId>::new();
+        for id in ids {
+            let cnt = locked
+                .ids
+                .remove(&id)
+                .expect("Attempted to unlock non-fungible that was not locked");
+            if cnt > 1 {
+                locked.ids.insert(id, cnt - 1);
+            } else {
+                liquid_non_fungibles.insert(id);
+            }
+        }
+
+        NonFungibleVault::put(
+            receiver,
+            LiquidNonFungibleResource::new(liquid_non_fungibles),
+            api,
+        )
     }
 }
